# backend/openapi/alpha.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bazaar internal Alpha API",
    "version": "0.1.0",
    "description": "Implemented internal Alpha contract. Pre-provisioned accounts only; simulated settlement. No uploads, registration, wallet or real payments. Every response returns X-Request-Id; errors contain code, retryable and the same requestId. Private/auth responses no-store."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/orders/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestId"
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "getOrder",
        "summary": "Participant only; outsiders and nonexistent IDs both 404",
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "400": {
            "description": "HTTP 400; code/retryable/requestId only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "401": {
            "description": "HTTP 401; code/retryable/requestId only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "404": {
            "description": "HTTP 404; code/retryable/requestId only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "500": {
            "description": "HTTP 500; code/retryable/requestId only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "description": "HTTP 503; code/retryable/requestId only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/admin/disputes": {
      "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
      "post": {
        "operationId": "listAdminDisputes",
        "summary": "Admin-only list of pending dispute decisions",
        "security": [{ "session": [], "csrf": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Empty" } } } },
        "responses": {
          "200": { "description": "Pending disputes", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminDisputePage" } } } },
          "default": { "description": "Error envelope", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/admin/disputes/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/RequestId" }, { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "post": {
        "operationId": "getAdminDispute",
        "summary": "Admin-only dispute evidence without delivery secrets",
        "security": [{ "session": [], "csrf": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Empty" } } } },
        "responses": {
          "200": { "description": "Dispute detail", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminDispute" } } } },
          "default": { "description": "Error envelope", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/admin/disputes/{id}/resolve": {
      "parameters": [{ "$ref": "#/components/parameters/RequestId" }, { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "post": {
        "operationId": "resolveAdminDispute",
        "summary": "Admin decides refund or release once, with simulated balance and audit updates",
        "security": [{ "session": [], "csrf": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DisputeResolutionInput" } } } },
        "responses": {
          "200": { "description": "Resolved order", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandResult" } } } },
          "default": { "description": "Error envelope", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "description": "Optional diagnostic ID. Invalid values are replaced; the effective ID is returned in the response header and error body.",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9_-]{1,100}$"
        }
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bazaar_session",
        "description": "HttpOnly SameSite=Lax Path=/; Secure outside loopback HTTP; fixed 12h."
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Obtained from login/session; required for authenticated writes in addition to exact Origin."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "retryable",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "retryable": {
            "type": "boolean"
          },
          "requestId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$"
          }
        }
      },
...
      "AdminDisputePage": {
        "type": "object", "additionalProperties": false, "required": ["items"],
        "properties": { "items": { "type": "array", "items": { "type": "object", "additionalProperties": false, "required": ["id", "status", "hasCounteroffer"], "properties": { "id": { "type": "string", "format": "uuid" }, "title": { "type": "string" }, "status": { "$ref": "#/components/schemas/OrderStatus" }, "currency": { "type": "string" }, "amount": { "type": "string" }, "hasCounteroffer": { "type": "boolean" } } } } }
      },
      "AdminDispute": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "status", "buyerId", "sellerId", "issue", "refundRequested", "counteroffer"],
        "properties": { "id": { "type": "string", "format": "uuid" }, "status": { "$ref": "#/components/schemas/OrderStatus" }, "buyerId": { "type": "string", "format": "uuid" }, "sellerId": { "type": "string", "format": "uuid" }, "title": { "type": "string" }, "type": { "$ref": "#/components/schemas/ListingType" }, "price": { "$ref": "#/components/schemas/Price" }, "issue": { "type": "string", "nullable": true }, "refundRequested": { "type": "boolean" }, "counteroffer": { "type": "string", "nullable": true } }
      },
      "DisputeResolutionInput": {
        "type": "object", "additionalProperties": false, "required": ["outcome"],
        "properties": { "outcome": { "type": "string", "enum": ["refund", "release"] }, "returnOutcome": { "type": "string", "enum": ["not_sent", "returned", "not_required"] } }
      },
...
      "Currency": {
        "type": "string",
        "enum": [
          "USD",
          "EUR",
          "GBP",
          "JPY",
          "CNY",
          "CAD",
          "AUD",
          "CHF",
          "HKD",
          "SGD",
          "KRW",
          "INR",
          "AED",
          "BRL",
          "BTC",
          "ETH",
          "USDT",
          "USDC",
          "SOL"
        ]
      },
      "Price": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "amount",
          "currency"
        ],
        "properties": {
          "amount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]{0,19})(\\.[0-9]{1,18})?$",
            "description": "Positive decimal string, at most 20 integer digits. Currency scales from GET /currencies; reject excess scale, never round."
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          }
        }
      },
      "ListingType": {
        "type": "string",
        "enum": [
          "physical",
          "digital"
        ]
      },
...
      "CommandResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "orderId"
        ],
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "pending_payment",
          "pending_delivery",
          "pending_acceptance",
          "issue",
          "completed",
          "cancelled",
          "expired",
          "refunded"
        ]
      },
...
      "Snapshot": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "listingVersion",
          "title",
          "description",
          "category",
          "type",
          "price",
          "licenseDescription",
          "contentVersion"
        ],
        "properties": {
          "listingVersion": {
            "type": "integer",
            "minimum": 1
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "type": {
            "$ref": "#/components/schemas/ListingType"
          },
          "price": {
            "$ref": "#/components/schemas/Price"
          },
          "licenseDescription": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "nullable": true
          },
          "contentVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "nullable": true
          }
        }
      },
      "Empty": {
        "type": "object",
        "additionalProperties": false,
        "properties": {}
      },
...
      "OrderDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "listingId",
          "buyerId",
          "sellerId",
          "status",
          "version",
          "title",
          "type",
          "price",
          "createdAt",
          "updatedAt",
          "snapshot",
          "shipping"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "listingId": {
            "type": "string",
            "format": "uuid"
          },
          "buyerId": {
            "type": "string",
            "format": "uuid"
          },
          "sellerId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "$ref": "#/components/schemas/ListingType"
          },
          "price": {
            "$ref": "#/components/schemas/Price"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "snapshot": {
            "$ref": "#/components/schemas/Snapshot"
          },
          "shipping": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "recipient",
              "contact",
              "address"
            ],
            "properties": {
              "recipient": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "contact": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "address": {
                "type": "string",
                "minLength": 1,
                "maxLength": 500
              }
            },
            "nullable": true
          }
        }
      },
...
    },
    "headers": {
      "RequestId": {
        "description": "Effective request correlation ID generated or accepted by the server.",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$"
        }
      }
    }
  }
}